Some refactoring of modules
