+++ lib/modules/components/ModuleIP.dart
@@ -1,0 +1,7 @@
+part of components;
+
+class ModuleIP extends PositionedModule {
+  ModuleIP(int id, String module, ModulePosition pos) : super(id, module, pos);
+  static instantiate(Map<String, dynamic> json) =>
+      PositionedModule.fromJson(json);
+}
