+++ lib/modules/components/components.dart
@@ -1,0 +1,16 @@
+library components;
+
+import 'package:smartmirror/modules/Module.dart';
+import 'package:flutter/material.dart';
+import 'package:smartmirror/modules/PositionedModule.dart';
+
+part "ModuleAlert.dart";
+part "ModuleCalendar.dart";
+part "ModuleClock.dart";
+part "ModuleCompliments.dart";
+part "ModuleForecast.dart";
+part "ModuleIP.dart";
+part "ModuleNewsfeed.dart";
+part "ModulePirSensor.dart";
+part "ModuleUpdateNotification.dart";
+part "ModuleWeather.dart";
